adjust row selector
